+++ src/commands/DbUpAction.php
@@ -10,10 +10,12 @@ use yii\console\ExitCode;
  *
  * @package fortrabbit\DeployTools\commands
  */
-class DbUpAction extends BaseAction
+class DbUpAction extends ConfigAwareBaseAction
 {
     /**
      * Upload database
+     *
+     * @param string|null $config Name of the deploy config
      *
      * @return int
      *
@@ -22,9 +24,8 @@ class DbUpAction extends BaseAction
      * @throws \fortrabbit\Copy\exceptions\PluginNotInstalledException
      * @throws \fortrabbit\Copy\exceptions\RemoteException
      * @throws \yii\base\Exception
-     * @throws \yii\console\Exception
      */
-    public function run()
+    public function run(string $config = null)
     {
         $plugin       = Plugin::getInstance();
         $path         = './storage/';
@@ -33,11 +34,19 @@ class DbUpAction extends BaseAction
         $steps        = 4;
         $messages     = [];
 
-        $app = $this->app ?: getenv(Plugin::ENV_NAME_APP);
+        $this->head(
+            "Export local DB and import on remote.",
+            "<comment>{$this->config}</comment> {$this->config->app}.frb.io",
+            $this->interactive ? true : false
+        );
 
-        $this->section("Export local DB - import on remote: $app");
+        // Always ask (default no), but skip question in non-interactive mode
+        if (!$this->confirm("Are you sure?", $this->interactive ? false : true)) {
+            return ExitCode::UNSPECIFIED_ERROR;
+        }
 
-        if (!$this->confirm("Are you sure?", true)) {
+        // Run 'before' commands and stop on error
+        if (!$this->runBeforeDeployCommands()) {
             return ExitCode::UNSPECIFIED_ERROR;
         }