+++ src/helpers/ArtisanStyleCommands.php
@@ -1,0 +1,18 @@
+<?php
+
+namespace fortrabbit\Copy\helpers;
+
+use ostark\Yii2ArtisanBridge\base\Commands;
+
+/**
+ * Class ArtisanStyleCommands
+ *
+ * @package fortrabbit\Copy\helpers
+ */
+class ArtisanStyleCommands extends Commands
+{
+    public function getHelpSummary()
+    {
+        return 'Copy Craft between environments.';
+    }
+}