+++ src/helpers/ConsoleOutputHelper.php
@@ -10,7 +10,7 @@ use Symfony\Component\Console\Helper\TableSeparator;
  *
  * @package fortrabbit\Copy\services
  *
- * @property string $app
+ * @property string  $app
  * @property boolean $dryRun
  */
 trait ConsoleOutputHelper
@@ -31,5 +31,33 @@ trait ConsoleOutputHelper
                 ['Dry run', $this->dryRun ? 'true' : 'false']
             ]
         );
+    }
+
+    /**
+     * Command line block
+     *
+     * @param string $cmd
+     *
+     * @return bool
+     */
+    public function cmdBlock(string $cmd)
+    {
+        $here = str_replace(getenv("HOME"), '~', getcwd());
+        $this->block($cmd, null, 'fg=white;bg=default', '<comment> ' . $here . ' ►  </comment>', false, false);
+        return true;
+    }
+
+    /**
+     * @param string      $message
+     * @param string|null $context
+     * @param bool        $clear
+     */
+    public function head(string $message, string $context = null, $clear = true)
+    {
+        if ($clear) {
+            $this->output->write(sprintf("\033\143"));
+        }
+
+        $this->block("<options=bold;fg=white>$message</>", $context, 'fg=white;', '▶ ', false, false);
     }
 }